# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MailUp - Backend Test",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Mostrar productos existentes",
                "description": "Permite detallar los productos registrados.",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "responses": {
                    "200": {
                        "description": "Operacion Exitosa."
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Registrar nuevo producto",
                "description": "Permite registrar un nuevo Producto.",
                "operationId": "acaa0d3fdc74fd4b534e70fc016cf7f1",
                "parameters": [
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Id de Marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Id de Categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Descripción",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Precio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price_sale",
                        "in": "query",
                        "description": "Precio Venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stock",
                        "in": "query",
                        "description": "Stock",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Imagen",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "422": {
                        "description": "El campo ya esta en uso."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener información de un Producto",
                "description": "Devuelve datos de un Producto específica según Id. ",
                "operationId": "249e765b6489191fb8167474db8867f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese Id de un Producto existente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa."
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    },
                    "404": {
                        "description": "Recurso no encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un Producto existente",
                "description": "Permite modificar datos de un Producto existente.",
                "operationId": "436a5b5d1aff1d5fd341389786161044",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de Producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Id de Marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Id de Categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Descripción",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Precio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price_sale",
                        "in": "query",
                        "description": "Precio Venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stock",
                        "in": "query",
                        "description": "Stock",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Imagen",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "422": {
                        "description": "El campo ya esta en uso."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar Producto existente",
                "description": "Permite eliminar un registro y no devuelve contenido.",
                "operationId": "cc01048629edd9759ba1d7ffc989f667",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de Producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa."
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    },
                    "404": {
                        "description": "Recurso no encontrado."
                    }
                }
            }
        }
    }
}